deprecated/requirements.txt
alembic==0.7.5.post2
beautifulsoup4==4.3.2
blinker==1.3
click==4.0
Flask-Cors==2.0.0
Flask-Limiter==0.7.9
flask-marshmallow==0.5.0
Flask-Migrate==1.3.1
flask-raven==0.0.4
Flask-Redis==0.0.6
Flask-RQ==0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
limits==1.0.4
Mako==1.0.1
MarkupSafe==0.23
marshmallow==1.2.4
MySQL-python==1.2.5
Pillow==2.8.1
pip-tools==0.3.6
psutil==2.2.1
psycopg2==2.6
pycrypto==2.6.1
PyPrind==2.9.1
pytz==2015.2
raven==5.2.0
redis==2.10.3
requests==2.6.0
rq==0.5.1
six==1.9.0
SQLAlchemy==1.0.0b5
Werkzeug==0.10.4
